refactor/#9 input 컴포넌트 리팩토링
- 컴포넌트 경로 변경
- 상위 컴포넌트 작업자가 직접 커스텀할 수 있도록 size 속성 제거
- state에 따른 색상 변경 속성 설정 기능 추가가
